Move `run_as_future` to `runtime`
